feat: add validation for user fields

+++ src/main/java/tech/ayot/ticket/backend/interceptor/RoleCheckerInterceptor.java
@@ -10,7 +10,7 @@ import org.springframework.web.servlet.HandlerInterceptor;
 import org.springframework.web.servlet.HandlerMapping;
 import tech.ayot.ticket.backend.annotation.CheckRole;
 import tech.ayot.ticket.backend.configuration.WebMvcConfiguration;
-import tech.ayot.ticket.backend.dto.auth.Role;
+import tech.ayot.ticket.backend.dto.auth.enumuration.Role;
 import tech.ayot.ticket.backend.service.auth.RoleService;
 
 import java.lang.reflect.Method;
@@ -57,7 +57,7 @@ public class RoleCheckerInterceptor implements HandlerInterceptor {
                 response.setStatus(HttpStatus.FORBIDDEN.value());
                 return false;
             }
-            int productId = Integer.parseInt(productIdString);
+            long productId = Integer.parseInt(productIdString);
 
             hasRole &= roleService.userHasRole(productId, checkRoleAnnotation.role());
         }
